backend/db/requirements.txt
aiohttp==3.7.3
aiosqlite==0.16.0
async-timeout==3.0.1
asyncpg==0.21.0
attrs==20.3.0
chardet==3.0.4
idna==2.10
iso8601==0.1.13
multidict==5.0.2
pkg-resources==0.0.0
pydantic==1.7.2
PyPika==0.44.0
pytz==2020.4
tortoise-orm==0.16.18
typing-extensions==3.7.4.3
yarl==1.6.3
